Refactor tests for suggest method to not hit real elasticsearch
We want to avoid relying on a real elasticsearch instance for unit tests in order to be able to make them faster and more repeatable. Instead create assertions on the queries that are passed into elastic search and the processing of the data it returns based on a mock search client.

Uses a fixture based on the result output by prod elastic earch at time of writing.