Exclude inactive tags from `suggest` query
Adds a filter to the elasticsearch query that explicitly excludes inactive tags. Note that the filter cannot require `active: true` because most of the json files in s3 do not have any active property at all. It is only tags which have recently been edited that will have the property present.