Fix AWS stubbing
If the stubbing is applied to the S3 prototype after an instance of S3 has been created then it fails due to the way AWS SDK loads classes (the *exact* cause of this is still unknown). By creating an instance and then stubbing *its* prototype method we can reliably stub actions on S3 instances.